Fix TypeError in calc_disk_usage function
The issue was caused by the calc_disk_usage function returning a string instead of an int when the path was not a directory. This resulted in a TypeError when trying to add the size of the subdirectories. The fix was to convert the return value to an int using the int() function.